Reduce service usage in AbstractExtensionService
